frontend/requirements.txt
streamlit>=1.20.0
requests>=2.28.0
pandas>=1.5.0
python-dateutil>=2.8.2
